"application, exception handler"

+++ src/main/java/br/com/fiap/abctechservice/model/OrderLocation.java
@@ -3,7 +3,7 @@ package br.com.fiap.abctechservice.model;
 import lombok.*;
 
 import javax.persistence.*;
-import java.sql.Date;
+import java.util.Date;
 
 @Entity
 @Table(name = "order_location")
@@ -18,9 +18,9 @@ public class OrderLocation {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Integer latitude;
+    private Double latitude;
 
-    private Integer longitude;
+    private Double longitude;
 
     private Date date;
